requirements.txt
python-telegram-bot==21.8
apscheduler==3.11.0
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.6
openai==1.54.0
